Create breakout
Here is my "Breakout" program. The biggest problem was "checking for collision". For this, I had to constantly check the x-y coordinates of the ball and see if it was either open, hitting a wall, or hitting a paddle. If it was hitting a wall or paddle, I just reversed the velocity like vy = -vy and vx = -vx.